+++ src/components/History/index.jsx
@@ -5,7 +5,7 @@ import Navbar from 'components/Navbar';
 import Button from 'components/Button';
 import Footer from 'components/Footer';
 import Banner from 'components/Banner';
-import HistoryTile from 'components/History/HistoryTile';
+import HistoryTiles from 'components/History/HistoryTiles';
 
 export default class History extends React.Component {
 	render() {
@@ -15,8 +15,8 @@ export default class History extends React.Component {
 				<Banner decorative />
 				<div className="content-section center">
 					<h2>History of ACM</h2>
-					<HistoryTile history={Config.history}/>
 				</div>
+				<HistoryTiles history={Config.history}/>
 				<Footer />
 			</div>
 		);
